add psr4, composer for autoload
