+++ composer.json
@@ -1,0 +1,5 @@
+{
+    "autoload": {
+        "psr-4": {"App\\": "src/"}
+    }
+}
