--- controllers/action.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace controllers;
-
-require_once('../models/Cat.php');
-require_once('../models/Dog.php');
-require_once('../models/Eagle.php');
-require_once('../models/Lizard.php');
-require_once('functions.php');
-
-use models\Cat;
-use models\Dog;
-use models\eagle;
-
-
-
-
-
-$dingo = new Dog("Dingo");
-$coco = new Eagle("Coco");
-$lizo = new \models\Lizard("lizo");
-$minette = new \models\Cat("Minette");
-
-
-
-
-
-
-
-
