--- models/Mamal.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace models;
-
-require_once('Animals.php');
-require_once('WalkingAnimal.php');
-
-class Mamal extends Animals implements walkingAnimal{
-
-    public function walking()
-    {
-        
-    }
-
-}
-
