--- models/WalkingAnimal.php
@@ -1,10 +1,0 @@
-<?php 
-
-namespace models;
-
-
-interface walkingAnimal{
-
-
-    public function walking();
-}
