+++ src/controllers/action.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Controllers;
+
+
+
+
+use App\Models\Dog;
+
+use App\Models\Eagle;
+use App\Models\Lizard;
+use App\Models\Cat;
+
+$dingo = new Dog("Dingo");
+$coco = new Eagle("Coco");
+$lizo = new Lizard("lizo");
+$minette = new Cat("Minette");
+
+
+
+
+
+
+
+
