+++ src/models/FlyingAnimals.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace models;
+namespace App\Models;
 
 interface FlyingAnimals{
