+++ src/models/Mamal.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace App\Models;
+
+
+class Mamal extends Animals implements WalkingAnimal{
+
+    public function walking()
+    {
+        
+    }
+
+}
+
