+++ src/models/WalkingAnimal.php
@@ -1,0 +1,10 @@
+<?php 
+
+namespace App\Models;
+
+
+interface WalkingAnimal{
+
+
+    public function walking();
+}
